complete refactor incoming
