--- phy/interleaver.py
@@ -1,52 +1,0 @@
-from typing import List, Union
-import numpy as np
-
-
-def interleaver(data: List[Union[int, complex]], N_bpcs: int) -> List[Union[int, complex]]:
-    """
-    Interleaver function based on 802.11n.
-
-    Parameters:
-    - data (List[Union[int, complex]]): input data to be interleaved.
-    - N_bpcs (int): number of coded bits per subcarrier.
-
-    Returns:
-    - List[Union[int, complex]]: interleaved data.
-    """
-    N_cbps = len(data)
-    num_columns = N_cbps // N_bpcs
-    interleaved_data: List[Union[int, complex]] = [0] * N_cbps
-
-    for k in range(N_cbps):
-        row = k % num_columns
-        column = k // num_columns
-
-        new_index = row * N_bpcs + column
-        interleaved_data[new_index] = data[k]
-
-    return interleaved_data
-
-
-def deinterleaver(interleaved_data: List[Union[int, complex]], N_bpcs: int) -> List[Union[int, complex]]:
-    """
-    Deinterleaver function based on 802.11n.
-
-    Parameters:
-    - interleaved_data (List[Union[int, complex]]): input data to be deinterleaved.
-    - N_bpcs (int): number of coded bits per subcarrier.
-
-    Returns:
-    - List[Union[int, complex]]: deinterleaved data.
-    """
-    N_cbps = len(interleaved_data)
-    num_columns = N_cbps // N_bpcs
-    deinterleaved_data: List[Union[int, complex]] = [0] * N_cbps
-
-    for k in range(N_cbps):
-        row = k % N_bpcs
-        column = k // N_bpcs
-
-        original_index = row * num_columns + column
-        deinterleaved_data[original_index] = interleaved_data[k]
-
-    return deinterleaved_data
